% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateCTriad.R
\name{calculateCTriad}
\alias{calculateCTriad}
\title{Calculate Conjoint Triad Descriptor}
\usage{
calculateCTriad(x)
}
\arguments{
\item{x}{A data.frame containing gene/protein names and their fasta sequences.}
}
\value{
A length 343  named vector for the data input.
}
\description{
This function calculates Conjoint Triad descriptor for data input.
}
\details{
calculateCTriad
}
\examples{
x <- readRDS(
    system.file("protseq/UP000464024_FASTAList.rds", package = "HPiP")
)
x_df <- calculateCTriad(x)
head(x_df, n = 2L)
}
\references{
Shen, J., Zhang, J., Luo, X., Zhu, W., Yu, K., Chen, K., et al.
(2007). Predicting protein–protein interactions based only on
sequences information. \emph{Proc. Natl. Acad. Sci.} 104, 4337–4341.
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
