% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateF1.R
\name{calculateF1}
\alias{calculateF1}
\title{Calculate Sum of Squared length of Single Amino Acid Repeats (SARs)}
\usage{
calculateF1(x)
}
\arguments{
\item{x}{A data.frame containing gene/protein names and their fasta sequences.}
}
\value{
A length 20 named vector for the data input.
}
\description{
This function calculates sum of quared length of Single Amino Acid Repeats (SARs)
in the entire protein sequence.
}
\details{
calculateF1
}
\examples{
x <- readRDS(
    system.file("protseq/UP000464024_FASTAList.rds", package = "HPiP")
)
x_df <- calculateF1(x)
head(x_df, n = 2L)
}
\references{
Alguwaizani, S., Park, B., Zhou, X., Huang, D.-S., and Han, K. (2018).
Predicting interactions between virus and host proteins using repeat patterns and composition of amino acids.
\emph{J. Healthc. Eng.} 2018.
}
\seealso{
See \code{\link{calculateF2}}
for maximum of the sum of squared length of SARs in a window of 6 residues descriptor.
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
