% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateF2.R
\name{calculateF2}
\alias{calculateF2}
\title{Calculate Maximum of the Sum of Squared Length of Single Amino Acid Repeats (SARs)}
\usage{
calculateF2(x)
}
\arguments{
\item{x}{A data.frame containing gene/protein names and their fasta sequences.}
}
\value{
A length 20 named vector for the data input.
}
\description{
This function calculates maximum of the sum of Single Amino Acid Repeats (SARs)
in a window of 6 residue.
}
\details{
calculateF2
}
\examples{
x <- readRDS(
    system.file("protseq/UP000464024_FASTAList.rds", package = "HPiP")
)
x_df <- calculateF2(x)
head(x_df, n = 2L)
}
\references{
Alguwaizani, S., Park, B., Zhou, X., Huang, D.-S., and Han, K. (2018).
Predicting interactions between virus and host proteins using repeat patterns and composition of amino acids.
\emph{J. Healthc. Eng.} 2018.
}
\seealso{
See \code{\link{calculateF1}}
for Sum of Squared length of Single Amino Acid Repeats (SARs) descriptor.
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
