% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.df.R
\docType{data}
\name{enrich.df}
\alias{enrich.df}
\title{Enrichment Result}
\usage{
data(enrich.df)
}
\description{
Input data for \code{\link[HPiP]{enrichplot}}
}
\details{
To construct this dataset, predicted interactions, generated
from \code{\link[HPiP]{pred_ensembel}} was used as data input for
\code{\link[HPiP]{enrichplot}}.
}
