% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichplot.R
\name{enrichplot}
\alias{enrichplot}
\title{Plot the Enrichment Reuslt}
\usage{
enrichplot(x, low = "blue", high = "red", cex.size = 15)
}
\arguments{
\item{x}{A data.frame with the enrichment analysis results.}

\item{low}{Colours for low.}

\item{high}{Colours for high.}

\item{cex.size}{Text size.}
}
\value{
An enrichment plot.
}
\description{
This function plots the enrichment result.
}
\details{
enrichplot
}
\examples{
enrich.result <- readRDS(system.file("examples/enirchment_result.rds", package = "HPiP"))
enrichplot(enrich.result,
    low = "blue",
    high = "red",
    cex.size = 12
)
}
\seealso{
See \code{\link{enrichfindP}} for functional enrichment analysis.
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
