% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Input Data for Prediction Algorithm}
\format{
a data.frame containing unlabeled or labeled HP-PPIs and pre-computed
numerical features.
}
\usage{
data(example_data)
}
\description{
Input data for \code{\link[HPiP]{pred_ensembel}}
}
