% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreqInteractors.R
\name{FreqInteractors}
\alias{FreqInteractors}
\title{Plot the Pathogen Proteins' Frequency of Interactions with Host Proteins}
\usage{
FreqInteractors(ppi, cex.size = 12)
}
\arguments{
\item{ppi}{A data.frame containing pathogen proteins in the first column and host
proteins in the second column.}

\item{cex.size}{Text size.}
}
\value{
A frequency plot.
}
\description{
This function plots the pathogen proteins' Frequency of interactions with host proteins
}
\details{
FreqInteractors
}
\examples{
ppi <- data.frame(
    node1 = c("A", "A", "A", "B", "B", "B", "B"),
    node2 = c("C", "E", "D", "F", "G", "H", "I")
)
FreqInteractors(ppi)
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
