% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FASTA.R
\name{get_FASTA}
\alias{get_FASTA}
\title{Retrieve FASTA Sequences from UniProt Databse}
\usage{
get_FASTA(custombg = NULL, proteomeIDs = "UP000464024")
}
\arguments{
\item{custombg}{A character vector of UniprotKB identifiers under study.}

\item{proteomeIDs}{A proteome identifier for the organism under study.The full list of
proteome IDs are available at  \url{https://www.uniprot.org/proteomes/}.}
}
\value{
A Data.frames containing protein FASTA sequences from the UniProt database.
}
\description{
This function retrieves FASTA sequences for the organism of interest
from the UniProt database and checks if the amino-acid composition of protein sequences
is in the 20 default types.
}
\details{
get_FASTA
}
\examples{
# get fasta seqeunces for three proteins of SARS-Cov-2
x <- c("P0DTC4", "P0DTC5", "P0DTC9")
fasta_df <- get_FASTA(custombg = x, proteomeIDs = "UP000464024")
head(fasta_df)
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
