% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host_se.R
\docType{data}
\name{host_se}
\alias{host_se}
\title{Host SummarizedExperiment object}
\usage{
data(host_se)
}
\description{
SummarizedExperiment object of numerical features for host
proteins.
}
\details{
To construct this object, first protein sequences were converted to
numerical features using (CTD) descriptors provided in the HPiP package,
followed by converting each numerical features matrix to
SummarizedExperiment object.Each object is then merged into one object
using `cbind()`.
}
