% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UP000464024_df.R
\docType{data}
\name{UP000464024_df}
\alias{UP000464024_df}
\title{Data.frame Containing SARS-CoV-2 FASTA Sequences}
\format{
A data.frame with two columns:(1) UniprotKBID, UniProt
identifier.(2) FASTASEQ, sequences per SARS-CoV-2 gene.
}
\source{
\url{https://www.uniprot.org/uniprot/?query=proteome:UP000464024}
}
\usage{
data(UP000464024_df)
}
\description{
This data includes one protein sequence per SARS-CoV-2 gene,
retrieved directly from UniProt database using
\code{\link[HPiP]{getFASTA}}.
}
