% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_imp.R
\name{var_imp}
\alias{var_imp}
\title{Variable Importance Plot}
\usage{
var_imp(x, cex.x = 1, cex.y = 2)
}
\arguments{
\item{x}{A list of elements returned from RFE analysis.
See \code{\link[caret]{rfe}} for more details}

\item{cex.x}{The size of x axis label.}

\item{cex.y}{The size of y axis label.}
}
\value{
Variable Importance Plot.
}
\description{
A graphical display of variable importance of selected features.
}
\details{
var_imp
}
\examples{
x <- readRDS(system.file("examples/training_data.rds", package = "HPiP"))
# select 50 rows
x <- x[1:50, 1:10]
# feature selection via rfe
x_output <- FSmethod(x,
    type = "rfe", resampling.method = "cv",
    iter = 2, repeats = NULL, metric = "Accuracy", verbose = TRUE
)
var_imp(x_output$rfProfile)
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}.
}
