% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viral_se.R
\docType{data}
\name{viral_se}
\alias{viral_se}
\title{Viral SummarizedExperiment object}
\usage{
data(viral_se)
}
\description{
SummarizedExperiment object of numerical features for
SARS-CoV-2 proteins.
}
\details{
To construct this object, first protein sequences were converted to
numerical features using (CTD) descriptors provided in the HPiP package,
followed by converting each numerical features matrix to
SummarizedExperiment object.
Each object is then merged into one object using `cbind()`.
}
