% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_positivePPI.R
\name{get_positivePPI}
\alias{get_positivePPI}
\title{Get Positive Reference Host-Pathogen Protein-Protein Interactions (HP-PPIs)}
\usage{
get_positivePPI(organism.taxID = 2697049)
}
\arguments{
\item{organism.taxID}{Taxonomy identifier for the pathogen. All identifiers are
available at \url{http://webservice.thebiogrid.org/organisms/?accessKey=81bb3b5a6bd9a8084a7be71f0963ab1e}.}
}
\value{
A Data.frame containing true positive protein-protein interactions for the selected
pathogen.
}
\description{
Get positive reference host-pathogen protein-protein interactions from BioGRID.
}
\details{
get_positivePPI
}
\examples{
positive_reflists <- get_positivePPI(organism.taxID = 2697049)
}
\seealso{
See \code{\link{get_negativePPI}} for generating negative
protein-protein interaction.
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
