% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_ensembel.R
\name{pred_ensembel}
\alias{pred_ensembel}
\title{Predict Interactions via Ensemble Learning Method}
\usage{
pred_ensembel(
  features,
  gold_standard,
  classifier = c("glm", "svmRadial", "rf"),
  resampling.method = "cv",
  ncross = 2,
  verboseIter = FALSE,
  plots = TRUE
)
}
\arguments{
\item{features}{A data frame with host-pathogen protein-protein interactions (HP-PPIs)
in the first column, and features to be passed to the classifier in the remaining columns.}

\item{gold_standard}{A data frame with gold_standard HP-PPIs and class label indicating
if such PPIs are positive or negative.}

\item{classifier}{The type of classifier to use. See \code{caret} for the
availbale classifiers.}

\item{resampling.method}{The resampling method:'boot', 'boot632', 'optimism_boot',
'boot_all', 'cv', 'repeatedcv', 'LOOCV', 'LGOCV'; defaults to cv. See \code{\link[caret]{trainControl}} for more details.}

\item{ncross}{Number of partitions for cross-validation;
defaults to 5.See \code{\link[caret]{trainControl}} for more details.}

\item{verboseIter}{Logical value, indicating whether to check the status of training process;
defaults to FALSE.}

\item{plots}{Logical value, indicating whether to plot the performance of ensemble learning algorithm
as compared to individual classifiers; defaults to TRUE.
If the argument set to TRUE, three plots will be generated:
\itemize{
\item{pr_curve} - precision-recall plot of ensemble classifier vs selected individual classifiers.
\item{roc_cruve} - ROC plot of ensemble classifier vs selected individual classifiers.
\item{Performance} - Plot accuracy, F1-score ,positive predictive value (PPV),sensitivity (SE),
and Matthews correlation coefficient (MCC) of ensemble classifier vs selected individual classifiers.
}}
}
\value{
Ensemble_training_output
\itemize{
\item{prediction score} - Prediction scores for whole dataset from each individual classifier.
\item{Best} - Selected hyper parameters.
\item{Parameter range} - Tested hyper parameters.
\item{prediction_score_test} - Scores probabilities for test data from each individual classifier.
\item{class_label} -  Class probabilities for test data from each individual classifier.
}
classifier_performance
\itemize{
\item{cm} - A confusion matrix.
\item{ACC} - Accuracy.
\item{SE} - Sensitivity.
\item{SP} - Specificity.
\item{PPV} - Positive Predictive Value.
\item{F1} - F1-score.
\item{MCC} - Matthews correlation coefficient.
\item{Roc_Object} - A list of elements. See \code{\link[pROC]{roc}} for more details.
\item{PR_Object} - A list of elements. See \code{\link[PRROC]{pr.curve}} for more details.
}
predicted_interactions - The input data frame of pairwise interactions, including
classifier scores averaged across all models.
}
\description{
This function uses an ensemble of classifiers to predict interactions from the
sequence-based dataset. This ensemble algorithm combines different results generated from individual
classifiers within the ensemble via average to enhance prediction.
}
\details{
pred_ensembel
}
\examples{
# load the independent data set
ind_data <- readRDS(system.file("examples/IndependentData_withFeatures.rds", package = "HPiP"))
# select important features
# load the gold-standard data set
gd <- readRDS(system.file("examples/Gold_ReferenceSet.rds", package = "HPiP"))
# select PPI and class label
gd <- gd[c(1, 5)]
ppi <- pred_ensembel(ind_data, gd,
    classifier = c("avNNet", "rf"),
    resampling.method = "cv",
    ncross = 5,
    verboseIter = FALSE,
    plots = FALSE
)
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
